requirements.txt
Flask
openai
python-dotenv
requests
mysql-connector-python
sounddevice
pyaudio
